+++ PreOrclBackEnd/Common.BOL/BOL/BOLUsuarios.cs
@@ -46,7 +46,7 @@ namespace Common.BOL.BOL
             return t;
         }
 
-        public Task<Usuarios> GetUsuario(int id) {
+        public Task<Usuarios> GetUsuario(decimal id) {
             Task<Usuarios> t = Task.Run(() =>
             {
                 using (DALDBContext context = new DALDBContext())
